refactor(tokens): break out AssetTabBar and AssetList into its own components
